Progress towards passing scene to Vulkan via SSBO and ray-tracing it. Camera and materials still buggy.

+++ compute_buffer.h
@@ -19,9 +19,9 @@ enum ComputeBufferVisibility {
 };
 
 struct ComputeBufferDims {
-    uint32_t width;
-    uint32_t height;
-    uint32_t elementSize;
+    size_t width;
+    size_t height;
+    size_t elementSize;
 
     ComputeBufferDims() = default;
 };
